+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.3
 
 - Feat: scan for unused definitions. Can be disabled by setting `Allow Unused Definitions` to `true`.
+- Feat: configurable update file debounce latency.
 - Fix: definition will be correctly refreshed if duplicate definition exists across files.
 - Fix: apply severity settings.